requirements.txt
requests
pandas
pandas-ta
numpy
mplfinance
python-telegram-bot
python-dotenv

